Move new interface to RoundDataTest
